@@ -14,13 +14,12 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.1 || ^8.0",
-        "ext-json": "*"
+        "php": "^8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.5.15 || ^8.5 || ^9.0",
+        "phpunit/phpunit": "^9.0",
         "php-coveralls/php-coveralls": "^2.2",
-        "vimeo/psalm": "4.9.2"
+        "vimeo/psalm": "5.12.0"
     },
     "autoload": {
         "psr-4": {
