@@ -6,9 +6,7 @@ namespace Brick\Math\Exception;
 
 /**
  * Base class for all math exceptions.
- *
- * This class is abstract to ensure that only fine-grained exceptions are thrown throughout the code.
  */
-class MathException extends \RuntimeException
+class MathException extends \Exception
 {
 }
